Module rtic_syntax::analyze

RTIC application analysis

Structs

Analysis
The result of analyzing an RTIC application

Channel
A channel between cores used to send messages

TimerQueue
The timer queue

Enums

Location
Resource location

Ownership
Resource ownership

Type Definitions

Ceiling
Priority ceiling

Channels
All cross-core channels, keyed by receiver core, then by dispatch priority and then by sender core

FreeQueues
All free queues, keyed by task and then by sender

InitializationBarriers
Cross-core initialization barriers

LateResources
Late resources, keyed by the core that initializes them

Locations
Location of all used resources

Ownerships
Resource ownership

Priority
Task priority

Receiver
Receiver core

Resource
Resource name

SendTypes
These types must implement the Send trait

Sender
Sender core

SpawnBarriers
Cross-core spawn barriers

SyncTypes
These types must implement the Sync trait

Task
Task name

TimerQueues
Timer queues, keyed by core