Struct rtic_syntax::analyze::Analysis

pub struct Analysis {
    pub used_cores: BTreeSet<Core>,
    pub channels: Channels,
    pub free_queues: FreeQueues,
    pub late_resources: LateResources,
    pub locations: Locations,
    pub ownerships: Ownerships,
    pub send_types: SendTypes,
    pub sync_types: SyncTypes,
    pub initialization_barriers: InitializationBarriers,
    pub spawn_barriers: SpawnBarriers,
    pub timer_queues: TimerQueues,
}

The result of analyzing an RTIC application

Fields
used_cores: BTreeSet<Core>

Cores that have been assigned at least task, #[init] or #[idle]

channels: Channels

SPSC message channels between cores

free_queues: FreeQueues

Priority ceilings of "free queues"

late_resources: LateResources

Maps a core to the late resources it initializes

locations: Locations

Location of all used resources

If a resource is not listed here it means that's a "dead" (never accessed) resource and the backend should not generate code for it

None indicates that the resource must reside in memory visible to more than one core ("shared memory")

ownerships: Ownerships

Resource ownership

send_types: SendTypes

These types must implement the Send trait

sync_types: SyncTypes

These types must implement the Sync trait

initialization_barriers: InitializationBarriers

Cross-core initialization barriers

spawn_barriers: SpawnBarriers

Cross-core spawn barriers

timer_queues: TimerQueues

Timer queues