Module rtic_syntax::ast

Abstract Syntax Tree

Structs

App
The #[app] attribute

AppArgs
The arguments of the #[app] attribute

ExternInterrupt
Interrupt that could be used to dispatch software tasks

HardwareTask
A hardware task

HardwareTaskArgs
Hardware task metadata

Idle
The idle context

IdleArgs
idle context metadata

Init
The init-ialization function

InitArgs
init context metadata

LateResource
A late (runtime initialized) resource

Local
A static mut variable local to and owned by a context

Resource
An early (compile time initialized) resource

SoftwareTask
A software task

SoftwareTaskArgs
Software task metadata

Enums

Access
Resource access

CustomArg
A custom argument

Type Definitions

ExternInterrupts
Interrupts used to dispatch software tasks

Idles
Per-core idle functions

Inits
Per-core init functions

Resources
Resource access list