rtic_syntax

Module ast

Abstract Syntax Tree

Structs

App

The #[app] attribute

AppArgs

The arguments of the #[app] attribute

ExternInterrupt

Interrupt that could be used to dispatch software tasks

HardwareTask

A hardware task

HardwareTaskArgs

Hardware task metadata

Idle

The idle context

IdleArgs

idle context metadata

Init

The init-ialization function

InitArgs

init context metadata

Local

A static mut variable local to and owned by a context

LocalResource

A local resource, defined in #[local]

Monotonic

Monotonic

MonotonicArgs

Monotonic metadata

SharedResource

A shared resource, defined in #[shared]

SharedResourceProperties

Shared resource properties

SoftwareTask

A software task

SoftwareTaskArgs

Software task metadata

Enums

Access

Resource access

TaskLocal

A wrapper of the 2 kinds of locals that tasks can have

Type Aliases

ExternInterrupts

Interrupts used to dispatch software tasks

LocalResources

Local resource access/declaration list in task attribute

SharedResources

Shared resource access list in task attribute